fixup! Add NoteEntity.kt

+++ app/src/main/java/com/laam/laamnotes/framework/db/NoteDao.kt
@@ -9,7 +9,7 @@ interface NoteDao {
     suspend fun addNoteEntity(noteEntity: NoteEntity)
 
     @Query("SELECT * FROM note WHERE id = :id")
-    suspend fun getNoteEntity(id: Int): NoteEntity?
+    suspend fun getNoteEntity(id: Long): NoteEntity?
 
     @Query("SELECT * FROM note")
     suspend fun getAllNoteEntity(): List<NoteEntity>
